Don't confuse a carried corpse for dragon-scaled armor.
Depending on the corpse number, this was causing an impossible(),
but it was wrong in any case:  carried corpses are not meant to
grant intrinsics.  The usual requirement that the item be worn
was not being handled correctly for corpses because comestibles
use the armor slot variable for something else.  The fix is to
check that the item is indeed armor, before checking whether
said armor is dragon-scaled.