Handle former-version Valkyrie player graves better.
In certain contexts (e.g., when making graves for the Valley of the Dead),
the game looks for lists of dead former players in historical records and
uses information about their role to create corpses.  Since Valkyrie is
not a playable role any more (because it's a race now), this logic was
failing to find the role in question (from its file code) in order to
get the appropriate permonst structure from a field in the roles array.

I solved this by inserting an explicit check for "Val".