When picking role for a new game, say how many roles remain locked.
Some players coming from vanilla or other variants were confused
when starting a new game, and thought it was a bug that not all
roles were instantly available the first time they ever played.
So, I finally dug into enough of the menu API to sort out how
to put some inactive text in this menu, to indicate that there
are additional role options that can be unlocked.