Don't abuse va_args and thus try to create strange objects.
My orchain implementation appears to have been incorrect, in a way
that resulted in its returning zero even when the last arg in the
list was definitely guaranteed to be nonzero.

In a player's tournament game, this was resulting in an attempt to
create a whole bunch of strange objects in a particular room, at
level generation time.

This, in turn, was resulting in a segfault (in principle, a series
of segfaults, if such a thing were possible), because if the player
chose to continue, the object was not created, and mksobj_at did not
take this possibility into account.  So I added a sanity check there,
to turn that segfault into an impossible.

And I ripped out the orchain and handled that a different way that
makes most of my code uglier but does not assign STRANGE_OBJECT
when it shouldn't.

I still can't successfully load the player's game, due to a log.c
error about not being able to upgrade to a write lock.  No idea
what that's about, but it seems unrelated, so I'm committing this
stuff before my brain goes catatonic for the night.