Officially start work toward 4.3.0.6
This is the first of the save-breaking changes after 4.3.0.5, and it
sets the version numbers for 4.3.0.6.

As with previous versions, my save-compatibility rule will be that
the master branch will remain save-compatible with 4.3.0.5 until
its version number changes, which will happen when a branch with
this commit on it gets merged back to master.  Meanwhile, the
bleeding-edge branch will already have this commit and may break
save compatibility at any time.

Public servers should generally use the master branch, and any
individuals who wish to play on the bleeding-edge branch should
do updates only between games, so as not to break their games in
progress.