The Great 4.3.0.6 Magic Cancellation Nerf
Armor no longer grants magic cancellation (except for the cloak of
protection, which grants it indirectly via the protection property).

Rings of protection still work, and I am contemplating an amulet of
protection.  Temporary protection is also available from the spell.

The problem with MC has always been that it does too many different
things, and in particular, it virtually eliminates the worst effects
of a whole bunch of different hazards.  Because players all just run
around with MC3 all the time, those various hazards that it (mostly)
protects against, have never really been balance tested outside of
the very early game, when a lot of them naturally do not occur.

I don't know if the MC nerf will be permanent or temporary, but I
know that I want to balance test what the game, and especially the
middle and late game, is like without it.  This may lead to who
knows what other balance changes.