When digging into stone, sometimes leave a boulder.
This serves several purposes.  It makes rocks more abundant, in
case a sling user does NOT want to use gold as ammo.  It makes
digging slightly more unpredictable and potentially interesting.
It serves as a source of boulders, for plugging up annoying holes
created by monsters with wands of digging.  And it makes sense
from a flavor perspective.